breaking: Move execution metadata API and remove automatic skill improvement
