+++ internal/runtime/discord/run.go
@@ -239,7 +239,7 @@ func run(ctx context.Context, b *Bot, in go_bot_discord.Input) error {
 	agentName := strings.TrimSpace(agent.Name())
 	sessionLog.Record(discordSessionID, agentTypes.Event{Type: agentTypes.EventAgentResult, Text: agentName})
 
-	execData := exec.ExecData{
+	execData := exec.ExecuteMeta{
 		Agent:          agent,
 		FallbackAgents: fallbacks,
 		WorkDir:        workDir,
